Fix incorrect use of scoped enumerations in format strings
Scoped enums do not get automatically promoted to their underlying type,
so these uses have undefined behavior and Clang recently started warning
about it.

Bug: 1456289
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4628112
Commit-Queue: Nico Weber <thakis@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Owners-Override: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1160036}

+++ ui/views/controls/menu/menu_controller_unittest.cc
@@ -498,9 +498,10 @@ class MenuControllerTest : public ViewsTestBase,
 
   void TestMenuFitsOnScreen(MenuAnchorPosition menu_anchor_position,
                             const gfx::Rect& monitor_bounds) {
-    SCOPED_TRACE(base::StringPrintf(
-        "MenuAnchorPosition: %d, monitor_bounds: @%s\n", menu_anchor_position,
-        monitor_bounds.ToString().c_str()));
+    SCOPED_TRACE(
+        base::StringPrintf("MenuAnchorPosition: %d, monitor_bounds: @%s\n",
+                           static_cast<int>(menu_anchor_position),
+                           monitor_bounds.ToString().c_str()));
     MenuBoundsOptions options;
     options.menu_anchor = menu_anchor_position;
     options.monitor_bounds = monitor_bounds;
@@ -547,9 +548,10 @@ class MenuControllerTest : public ViewsTestBase,
 
   void TestMenuFitsOnScreenSmallAnchor(MenuAnchorPosition menu_anchor_position,
                                        const gfx::Rect& monitor_bounds) {
-    SCOPED_TRACE(base::StringPrintf(
-        "MenuAnchorPosition: %d, monitor_bounds: @%s\n", menu_anchor_position,
-        monitor_bounds.ToString().c_str()));
+    SCOPED_TRACE(
+        base::StringPrintf("MenuAnchorPosition: %d, monitor_bounds: @%s\n",
+                           static_cast<int>(menu_anchor_position),
+                           monitor_bounds.ToString().c_str()));
     MenuBoundsOptions options;
     options.menu_anchor = menu_anchor_position;
     options.monitor_bounds = monitor_bounds;
@@ -599,9 +601,10 @@ class MenuControllerTest : public ViewsTestBase,
 
   void TestMenuFitsOnSmallScreen(MenuAnchorPosition menu_anchor_position,
                                  const gfx::Rect& monitor_bounds) {
-    SCOPED_TRACE(base::StringPrintf(
-        "MenuAnchorPosition: %d, monitor_bounds: @%s\n", menu_anchor_position,
-        monitor_bounds.ToString().c_str()));
+    SCOPED_TRACE(
+        base::StringPrintf("MenuAnchorPosition: %d, monitor_bounds: @%s\n",
+                           static_cast<int>(menu_anchor_position),
+                           monitor_bounds.ToString().c_str()));
     MenuBoundsOptions options;
     options.menu_anchor = menu_anchor_position;
     options.monitor_bounds = monitor_bounds;
